Add in prompt for itch-project name.
We now ask for the project name on itch.io if the user provides a
username for itch.io. This is optional, and if left-blank we
fallback to project-name.

I have to do this weirdly because I can't use template
placeholders in cargo-generate.toml for defaults, as far as I
could tell. We *could* create some hooks, but that is complicated
and there are some issues we could run-into so I went this route
instead so we wouldn't have to worry as much about it.