Add pause menu (#419)
Relates to https://github.com/TheBevyFlock/bevy_new_2d/issues/206.

NOTES:
- This PR only adds the menu without actually pausing gameplay.
- This PR doesn't add a translucent pause overlay.
- I hard-coded the "go back" feature, which would be better implemented
with a state stack.
- I can't write "upon exiting any screen, close the current menu" using
`bevy_state` because there is no state pattern-matching, so this is
hard-coded as well. (well I can, but it wouldn't be simple enough for
this template)
- There are probably 1-frame delays between entering a `Screen` and
opening / closing the relevant `Menu`, because I can't do dependent
states in `bevy_state` (aka require state transitions of `Menu` to run
`.after` state transitions of `Screen` so that the latter can control
the former on transition within the same frame).


![2025-05-23_1748051220_1920x1038](https://github.com/user-attachments/assets/65500fe0-560b-4b0c-bc9c-5f729ba3405c)


![2025-05-23_1748051200_1920x1038](https://github.com/user-attachments/assets/60ba8e60-c841-40ea-b36f-a8d15c250cc8)